use numpy in trigen

+++ declafe/feature_gen/tri/talib/ADXRFeature.py
@@ -13,7 +13,7 @@ class ADXRFeature(TriFeature):
 
   def trigen(self, col1: pd.Series, col2: pd.Series,
              col3: pd.Series) -> pd.Series:
-    return talib.ADXR(col1, col2, col3, self.period)
+    return talib.ADXR(col1.astype(float), col2.astype(float), col3.astype(float), self.period)
 
   def _feature_name(self) -> str:
     return f"ADXR_{self.period}_of_{self.col1}_{self.col2}_{self.col3}"
