+++ declafe/feature_gen/tri/talib/CCIFeature.py
@@ -16,7 +16,7 @@ class CCIFeature(TriFeature):
 
   def trigen(self, col1: pd.Series, col2: pd.Series,
              col3: pd.Series) -> pd.Series:
-    return talib.CCI(col1, col2, col3, self.period)
+    return talib.CCI(col1.astype(float), col2.astype(float), col3.astype(float), self.period)
 
   def _feature_name(self) -> str:
     return f"CCI_{self.period}_of_{self.col1}_{self.col2}_{self.col3}"
