+++ declafe/feature_gen/tri/talib/TRANGEFeature.py
@@ -17,7 +17,7 @@ class TRANGEFeature(TriFeature):
 
   def trigen(self, col1: pd.Series, col2: pd.Series,
              col3: pd.Series) -> pd.Series:
-    return talib.TRANGE(col1, col2, col3)
+    return talib.TRANGE(col1.astype(float), col2.astype(float), col3.astype(float))
 
   def _feature_name(self) -> str:
     return f"TRANGE_{self.col1}_{self.col2}_{self.col3}"
