+++ declafe/feature_gen/tri/talib/ULTOSCFeature.py
@@ -16,9 +16,9 @@ class ULTOSCFeature(TriFeature):
 
   def trigen(self, col1: pd.Series, col2: pd.Series,
              col3: pd.Series) -> pd.Series:
-    return talib.ULTOSC(col1,
-                        col2,
-                        col3,
+    return talib.ULTOSC(col1.astype(float),
+                        col2.astype(float),
+                        col3.astype(float),
                         timeperiod1=self.timeperiod1,
                         timeperiod2=self.timeperiod2,
                         timeperiod3=self.timeperiod3)
